requirements.txt
absl-py==0.8.1
appnope==0.1.0
asn1crypto==0.24.0
astor==0.8.0
atari-py==0.2.6
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.2
chardet==3.0.4
cloudpickle==1.2.2
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gym==0.15.4
h5py==2.9.0
idna==2.8
ignite==1.1.0
imageio==2.6.1
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter-client==5.3.3
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
musthe==1.0.0
nbconvert==5.6.0
nbformat==4.4.0
nbopen==0.6
notebook==6.0.1
numpy==1.17.2
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycosat==0.6.3
pycparser==2.19
pyglet==1.3.2
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.15.4
PySocks==1.6.8
python-dateutil==2.8.0
pytorch-lightning==0.5.3.2
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.1.0
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-learn==0.20.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn-sequence-classifiers==0.2
soupsieve==1.9.1
stable-baselines==2.8.0
tensorboard==2.0.1
tensorboardX==1.9
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.2
test-tube==0.7.5
testpath==0.4.2
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.35.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.11.2
